fix: change dialog to refer to document inside useEffect
